# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Event Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  participant: 'participant',
  organizer: 'organizer',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  organizer: 'organizer',
  participant: 'participant',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
  isActive?: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
} as const;

export type EventApprovalStatus = typeof EventApprovalStatus[keyof typeof EventApprovalStatus];


export const EventApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  venue: string;
  category: string;
  capacity: number;
  registrationDeadline: string;
  status: EventStatus;
  approvalStatus: EventApprovalStatus;
  /** @nullable */
  bannerUrl?: string | null;
  organizerId: number;
  /** @nullable */
  organizerName?: string | null;
  registrationCount?: number;
  isRegistered?: boolean;
  createdAt: string;
}

export type EventDetailStatus = typeof EventDetailStatus[keyof typeof EventDetailStatus];


export const EventDetailStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
} as const;

export type EventDetailApprovalStatus = typeof EventDetailApprovalStatus[keyof typeof EventDetailApprovalStatus];


export const EventDetailApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface EventDetail {
  id: number;
  title: string;
  description: string;
  date: string;
  venue: string;
  category: string;
  capacity: number;
  registrationDeadline: string;
  status: EventDetailStatus;
  approvalStatus: EventDetailApprovalStatus;
  /** @nullable */
  bannerUrl?: string | null;
  organizerId: number;
  /** @nullable */
  organizerName?: string | null;
  /** @nullable */
  organizerEmail?: string | null;
  registrationCount?: number;
  isRegistered?: boolean;
  /** @nullable */
  registrationId?: number | null;
  createdAt: string;
}

export interface EventInput {
  title: string;
  description: string;
  date: string;
  venue: string;
  category: string;
  capacity: number;
  registrationDeadline: string;
  bannerUrl?: string;
}

export interface EventUpdate {
  title?: string;
  description?: string;
  date?: string;
  venue?: string;
  category?: string;
  capacity?: number;
  registrationDeadline?: string;
  bannerUrl?: string;
  status?: string;
}

export type EventStatusUpdateApprovalStatus = typeof EventStatusUpdateApprovalStatus[keyof typeof EventStatusUpdateApprovalStatus];


export const EventStatusUpdateApprovalStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface EventStatusUpdate {
  approvalStatus: EventStatusUpdateApprovalStatus;
  rejectionReason?: string;
}

export interface EventListResponse {
  events: Event[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Attendee {
  registrationId: number;
  userId: number;
  name: string;
  email: string;
  registeredAt: string;
  attended: boolean;
}

export interface AttendeeListResponse {
  attendees: Attendee[];
  total: number;
  page: number;
  totalPages: number;
}

export interface DailyCount {
  date: string;
  count: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface EventAnalytics {
  eventId: number;
  totalRegistrations: number;
  capacity: number;
  attendanceRate: number;
  registrationsByDay: DailyCount[];
  categoryBreakdown?: CategoryCount[];
}

export interface Registration {
  id: number;
  eventId: number;
  userId: number;
  registeredAt: string;
  attended: boolean;
  /** @nullable */
  qrCode?: string | null;
}

export interface RegistrationWithEvent {
  id: number;
  eventId: number;
  userId: number;
  registeredAt: string;
  attended: boolean;
  /** @nullable */
  qrCode?: string | null;
  event?: Event;
}

export interface RegistrationListResponse {
  registrations: RegistrationWithEvent[];
  total: number;
  page: number;
  totalPages: number;
}

export interface QrCodeResponse {
  qrCode: string;
  registrationId: number;
  eventTitle: string;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalEvents: number;
  activeRegistrations: number;
  upcomingEvents: number;
  pendingApprovals?: number;
  eventsByCategory: CategoryCount[];
  recentActivity: ActivityItem[];
}

export interface EventSummary {
  id: number;
  title: string;
  registrationCount: number;
  capacity: number;
}

export interface OrganizerStats {
  totalEvents: number;
  totalRegistrations: number;
  upcomingEvents: number;
  completedEvents: number;
  pendingApproval?: number;
  topEvents?: EventSummary[];
}

export type ListUsersParams = {
page?: number;
limit?: number;
role?: string;
search?: string;
};

export type ListEventsParams = {
page?: number;
limit?: number;
category?: string;
status?: string;
location?: string;
search?: string;
dateFrom?: string;
dateTo?: string;
sortBy?: string;
sortOrder?: string;
};

export type GetMyRegistrationsParams = {
page?: number;
limit?: number;
};