# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}